#23 move read operations to collections subfolder
